+++ CHANGELOG.md
@@ -16,6 +16,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - `ShortestPathRouter` router pass
 - `RandomMapper` mapper pass
+- `AStarRouter` router pass
 
 ### Changed